// Angular/Avancado/angular-filmes/src/app/shared/models/filme.ts
export interface Filme {
  id?: number;
  titulo: string;
  urlFoto?: string;
  dtLancamento: Date;
  descricao?: string;
  nota: number;
  urlIMDb: string;
  genero: string;
}

export interface Gama {
  id?: number;
  titulo: string;
  urlFoto?: string;
  dtLancamento: Date;
  descricao?: string;
  nota: number;
  urlIMDb: string;
  genero: string;
}